+++ test/run_test.py
@@ -45,7 +45,8 @@ TESTS = [
     'distributed/test_c10d_gloo',
     'distributed/test_c10d_nccl',
     'distributed/test_jit_c10d',
-    'distributed/test_c10d_spawn',
+    'distributed/test_c10d_spawn_gloo',
+    'distributed/test_c10d_spawn_nccl',
     'test_cuda',
     'test_jit_cuda_fuser',
     'test_cuda_primary_ctx',
@@ -290,7 +291,8 @@ TARGET_DET_LIST = [
     'distributed/test_c10d_gloo',
     'distributed/test_c10d_nccl',
     'distributed/test_jit_c10d',
-    'distributed/test_c10d_spawn',
+    'distributed/test_c10d_spawn_gloo',
+    'distributed/test_c10d_spawn_nccl',
     'test_quantization',
     'test_pruning_op',
     'test_determination',
